tools/pyproject.toml
[project]
name = "tools"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.27.2",
    "jinja2>=3.1.5",
    "pyyaml>=6.0.2",
]

[dependency-groups]
dev = [
    "ruff>=0.8.4",
]

[tool.ruff]
line-length = 100
